requirements.txt
Click
fastapi
h11
httptools
psycopg2-binary
pydantic
SQLAlchemy
starlette
typing
uvicorn
uvloop
websockets